разбить утилитарный модуль на два модуля - общий и утилиты для фильмов
