+++ src/utils/film.js
@@ -10,15 +10,6 @@ const convertMinutesToHours = (minutesDuration) => {
   return hours < 1 ? `${minutes}m` : `${hours}h ${minutes}m`;
 };
 
-const getRandomInteger = (a = 0, b = 1) => {
-  const lower = Math.ceil(Math.min(a, b));
-  const upper = Math.floor(Math.max(a, b));
-
-  return Math.floor(lower + Math.random() * (upper - lower + 1));
-};
-
-const getRandomItem = (items) => items[getRandomInteger(0, items.length - 1)];
-
 const formatISOStringToDate = (ISOString) => dayjs(ISOString).format('DD MMMM YYYY');
 
 const formatISOStringToDateWithTime = (ISOString) => dayjs(ISOString).format('YYYY/MM/DD HH:MM');
@@ -29,8 +20,6 @@ const formatMinutesToTime = (minutes) => dayjs.duration(minutes, 'minutes').form
 
 export {
   convertMinutesToHours,
-  getRandomInteger,
-  getRandomItem,
   formatISOStringToDate,
   formatISOStringToYear,
   formatMinutesToTime,
